♻️ refactor: atomic 패턴제외하기위해 우선 ui라는 폴더명으로 변경 #70
